Cache results of API calls
The Weather API is rate limited so this is a quick way to cache the results of the API call so it is only called once per application run.

Refreshing the site will throw away this cache so the cache could be improved a lot by storing it in local storage in the future.